+++ src/widgets/FeedbackList/useFeedbackList.ts
@@ -1,0 +1,39 @@
+import { useEffect, useMemo, useState } from 'react';
+
+import { TFeedbackItemProps } from '../../shared/components/FeedbackItem/type';
+
+interface UseFeedbackListProps {
+  initialData: TFeedbackItemProps[];
+  onDataChange: (feedbacks: TFeedbackItemProps[]) => void;
+}
+
+const useFeedbackList = ({
+  initialData,
+  onDataChange
+}: UseFeedbackListProps) => {
+  const [feedbacks, setFeedbacks] = useState<TFeedbackItemProps[]>(initialData);
+  const [isAscending, setIsAscending] = useState<boolean>(true);
+
+  useEffect(() => {
+    onDataChange(feedbacks);
+  }, [feedbacks, onDataChange]);
+
+  const sortedFeedbacks = useMemo(() => {
+    return [...feedbacks].sort((a, b) => {
+      return isAscending
+        ? a.date.getTime() - b.date.getTime()
+        : b.date.getTime() - a.date.getTime();
+    });
+  }, [feedbacks, isAscending]);
+
+  const toggleSort = () => setIsAscending((prev) => !prev);
+
+  return {
+    sortedFeedbacks,
+    setFeedbacks,
+    isAscending,
+    toggleSort
+  };
+};
+
+export default useFeedbackList;
